backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]
passlib[bcrypt]
python-multipart==0.0.6
bcrypt==4.0.1 